refactor(cache.sqlite): move invalid route exports to `cache.server.ts`
