+++ app/routes/resources+/cache.sqlite.ts
@@ -1,8 +1,5 @@
 import {json, redirect, type DataFunctionArgs} from '@remix-run/node'
-import {
-  getInstanceInfo,
-  getInternalInstanceDomain,
-} from '~/utils/cjs/litefs-js.server.js'
+import {getInstanceInfo} from '~/utils/cjs/litefs-js.server.js'
 import {cache} from '~/utils/cache.server.ts'
 import {getRequiredServerEnvVar} from '~/utils/misc.tsx'
 
@@ -32,29 +29,4 @@ export async function action({request}: DataFunctionArgs) {
     await cache.set(key, cacheValue)
   }
   return json({success: true})
-}
-
-export async function updatePrimaryCacheValue({
-  key,
-  cacheValue,
-}: {
-  key: string
-  cacheValue: any
-}) {
-  const {currentIsPrimary, primaryInstance} = await getInstanceInfo()
-  if (currentIsPrimary) {
-    throw new Error(
-      `updatePrimaryCacheValue should not be called on the primary instance (${primaryInstance})}`,
-    )
-  }
-  const domain = getInternalInstanceDomain(primaryInstance)
-  const token = getRequiredServerEnvVar('INTERNAL_COMMAND_TOKEN')
-  return fetch(`${domain}/resources/cache/sqlite`, {
-    method: 'POST',
-    headers: {
-      Authorization: `Bearer ${token}`,
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({key, cacheValue}),
-  })
 }
